+++ crates/autopilot/src/boundary/order.rs
@@ -37,45 +37,3 @@ pub fn to_domain(
         signature: order.signature.into(),
     }
 }
-
-/// Recover order uid from order data and signature
-pub fn order_uid(
-    trade: &domain::settlement::coded::tokenized::Trade,
-    tokens: &[domain::settlement::coded::tokenized::Token],
-    domain_separator: &domain::eth::DomainSeparator,
-) -> Result<domain::OrderUid, Error> {
-    let flags = domain::settlement::coded::TradeFlags(trade.8);
-    let signature = crate::boundary::Signature::from_bytes(flags.signing_scheme(), &trade.10 .0)
-        .map_err(Error::Signature)?;
-
-    let order = model::order::OrderData {
-        sell_token: tokens[trade.0.as_u64() as usize],
-        buy_token: tokens[trade.1.as_u64() as usize],
-        sell_amount: trade.3,
-        buy_amount: trade.4,
-        valid_to: trade.5,
-        app_data: crate::boundary::AppDataHash(trade.6 .0),
-        fee_amount: trade.7,
-        kind: match flags.side() {
-            domain::auction::order::Side::Buy => model::order::OrderKind::Buy,
-            domain::auction::order::Side::Sell => model::order::OrderKind::Sell,
-        },
-        partially_fillable: flags.partially_fillable(),
-        receiver: Some(trade.2),
-        sell_token_balance: flags.sell_token_balance(),
-        buy_token_balance: flags.buy_token_balance(),
-    };
-    let domain_separator = crate::boundary::DomainSeparator(domain_separator.0);
-    let owner = signature
-        .recover_owner(&trade.10 .0, &domain_separator, &order.hash_struct())
-        .map_err(Error::RecoverOwner)?;
-    Ok(order.uid(&domain_separator, &owner).into())
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error("bad signature {0}")]
-    Signature(anyhow::Error),
-    #[error("recover owner {0}")]
-    RecoverOwner(anyhow::Error),
-}
